--- ToolMenu/submenus/toolsmenu.py
@@ -1,12 +1,0 @@
-import tkinter as tk
-
-
-class ToolsMenu(tk.Frame):
-    def __init__(self, parent, controller):
-        tk.Frame.__init__(self, parent)
-        label = tk.Label(self, text="Tools Menu")
-        label.pack(pady=10, padx=10)
-
-        button1 = tk.Button(self, text="Back to Home",
-                            command=lambda: controller.show_frame(parent.MainMenu))
-        button1.pack()
